`MACs/cmac` benchmarks were used to evaluate the refactoring, as `dbl` does not contain its own. The refactoring of `dbl` seems to have no effect on the `MACs/cmac` performance.

Benchmarks results with `dbl` refactored:
```
$ git checkout 90c46e717ab2321022104ead2f9060b1327f5710; rustup run nightly cargo bench
Previous HEAD position was 9d876cb Bump digest from 0.10.0 to 0.10.1 (#100)
HEAD is now at 90c46e7 Test dbl safe code
    Finished bench [optimized] target(s) in 0.02s
     Running unittests (/home/alwagner/working_directory/tests/MACs/target/release/deps/cmac-5a75094f2021ff95)

running 0 tests

test result: ok. 0 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; finished in 0.00s

     Running unittests (/home/alwagner/working_directory/tests/MACs/target/release/deps/aes128_cmac-7ce19bbcc576eef1)

running 4 tests
test bench1_10    ... bench:         235 ns/iter (+/- 102) = 42 MB/s
test bench2_100   ... bench:       2,294 ns/iter (+/- 55) = 43 MB/s
test bench3_1000  ... bench:      22,824 ns/iter (+/- 1,547) = 43 MB/s
test bench3_10000 ... bench:     228,810 ns/iter (+/- 16,931) = 43 MB/s

test result: ok. 0 passed; 0 failed; 0 ignored; 4 measured; 0 filtered out; finished in 5.51s

     Running unittests (/home/alwagner/working_directory/tests/MACs/target/release/deps/aes256_cmac-e6b672d2c21213f3)

running 4 tests
test bench1_10    ... bench:         312 ns/iter (+/- 10) = 32 MB/s
test bench2_100   ... bench:       3,030 ns/iter (+/- 159) = 33 MB/s
test bench3_1000  ... bench:      30,607 ns/iter (+/- 2,214) = 32 MB/s
test bench3_10000 ... bench:     303,722 ns/iter (+/- 16,212) = 32 MB/s

test result: ok. 0 passed; 0 failed; 0 ignored; 4 measured; 0 filtered out; finished in 3.86s
```

Benchmark results with `dbl` untouched:
```
$ git checkout 9d876cb237bbbddee527d8126663268bea91b6e9; rustup run nightly cargo bench
Previous HEAD position was 90c46e7 Test dbl safe code
HEAD is now at 9d876cb Bump digest from 0.10.0 to 0.10.1 (#100)
    Finished bench [optimized] target(s) in 0.02s
     Running unittests (/home/alwagner/working_directory/tests/MACs/target/release/deps/cmac-b1ec60cee9662e1a)

running 0 tests

test result: ok. 0 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; finished in 0.00s

     Running unittests (/home/alwagner/working_directory/tests/MACs/target/release/deps/aes128_cmac-137d219e2208c539)

running 4 tests
test bench1_10    ... bench:         236 ns/iter (+/- 13) = 42 MB/s
test bench2_100   ... bench:       2,286 ns/iter (+/- 124) = 43 MB/s
test bench3_1000  ... bench:      22,978 ns/iter (+/- 1,822) = 43 MB/s
test bench3_10000 ... bench:     228,643 ns/iter (+/- 10,610) = 43 MB/s

test result: ok. 0 passed; 0 failed; 0 ignored; 4 measured; 0 filtered out; finished in 1.73s

     Running unittests (/home/alwagner/working_directory/tests/MACs/target/release/deps/aes256_cmac-b77e6ae6be42117d)

running 4 tests
test bench1_10    ... bench:         312 ns/iter (+/- 54) = 32 MB/s
test bench2_100   ... bench:       3,070 ns/iter (+/- 128) = 32 MB/s
test bench3_1000  ... bench:      30,668 ns/iter (+/- 2,124) = 32 MB/s
test bench3_10000 ... bench:     304,873 ns/iter (+/- 15,266) = 32 MB/s

test result: ok. 0 passed; 0 failed; 0 ignored; 4 measured; 0 filtered out; finished in 12.60s
```