@aewag 2021-12-20
> Thank you! It would be nice to get performance measurements without the `force-soft` feature enabled for `aes`, but I guess performance should be the same.

Below is the benchmark without `force-soft` feature. PR seems to have no effect.

Before:
```
running 4 tests                                     
test bench1_10    ... bench:          19 ns/iter (+/- 1) = 526 MB/s
test bench2_100   ... bench:         118 ns/iter (+/- 6) = 847 MB/s
test bench3_1000  ... bench:       1,097 ns/iter (+/- 42) = 911 MB/s
test bench3_10000 ... bench:      10,927 ns/iter (+/- 455) = 915 MB/s
                                                    
running 4 tests                                                                                          
test bench1_10    ... bench:          22 ns/iter (+/- 2) = 454 MB/s
test bench2_100   ... bench:         148 ns/iter (+/- 9) = 675 MB/s                                      
test bench3_1000  ... bench:       1,412 ns/iter (+/- 45) = 708 MB/s
test bench3_10000 ... bench:      13,912 ns/iter (+/- 597) = 718 MB/s
```

After (`dbl` refactored):
```
running 4 tests                                                                                          
test bench1_10    ... bench:          19 ns/iter (+/- 3) = 526 MB/s        
test bench2_100   ... bench:         117 ns/iter (+/- 4) = 854 MB/s 
test bench3_1000  ... bench:       1,115 ns/iter (+/- 34) = 896 MB/s                   
test bench3_10000 ... bench:      10,960 ns/iter (+/- 258) = 912 MB/s

running 4 tests
test bench1_10    ... bench:          21 ns/iter (+/- 1) = 476 MB/s
test bench2_100   ... bench:         146 ns/iter (+/- 7) = 684 MB/s
test bench3_1000  ... bench:       1,411 ns/iter (+/- 94) = 708 MB/s
test bench3_10000 ... bench:      14,091 ns/iter (+/- 360) = 709 MB/s
```
> While we at it can you please also add #[inline] attributes on methods?

Done.